Added first version of run file and clean-up script. Have done:
Checking of availability of necessary commands.
Integrated creation of the necessary pg docker + odoo docker to be able to run the tests.
Added parsing of output, and putting up clear message on the screen
Detecting changes to the module being monitored using a combination of inotifwait + md5sum hashing

To do:
Add parameter to allow specifying which module(s) to test.
Adding parameter/config option to allow selection of odoo and pg image.